move surface area class into area class
