+++ maths/area.py
@@ -1,8 +1,33 @@
 """
 Find the area of various geometric shapes
 """
-
 import math
+
+
+def surface_area_cube(side_length: Union[int, float]) -> float:
+    """
+    Calculate the Surface Area of a Cube.
+
+    >>> surface_area_cube(1)
+    6.0
+    >>> surface_area_cube(3)
+    54.0
+    """
+    return 6 * pow(side_length, 2)
+
+
+def surface_area_sphere(radius: float) -> float:
+    """
+    Calculate the Surface Area of a Sphere.
+    Wikipedia reference: https://en.wikipedia.org/wiki/Sphere
+    :return 4 * pi * r^2
+
+    >>> vol_sphere(5)
+    314.1592653589793
+    >>> vol_sphere(1)
+    12.566370614359172
+    """
+    return 4 * pi * pow(radius, 2)
 
 
 def area_rectangle(base, height):
@@ -73,7 +98,9 @@ def main():
     print(f"Parallelogram: {area_parallelogram(10, 20)=}")
     print(f"Trapezium: {area_trapezium(10, 20, 30)=}")
     print(f"Circle: {area_circle(20)=}")
-
+    print("Surface Areas of various geometric shapes: \n")
+    print(f"Cube: {surface_area_cube(20)=}")
+    print(f"Sphere: {surface_area_sphere(20)=}")
 
 if __name__ == "__main__":
     main()
